refactor: make code more clean by using a barrel exports to break up the code in server.ts
