+++ src/server/index.ts
@@ -1,0 +1,3 @@
+export { server } from './server'
+
+
