+++ src/server/server.ts
@@ -1,13 +1,14 @@
 import express, { Application } from "express";
-import { loginRouter } from './routes/login'
-
+import { loginRouter } from '../routes/index'
 import * as dotenv from 'dotenv'
+
 dotenv.config({path: '.env'});
-export const { PORT,CLIENT_ID,CLIENT_SECERET } = process.env;
 
-
+export const { PORT } = process.env;
 export const server: Application = express();
-server.use('/login', loginRouter)
+
+
+server.use(loginRouter)
 
 
 if (process.env['NODE_ENV'] !== 'test') {
